feat(settings): Save is the Hub's only primary button
Settings fixtures move to conftest.py so the new test files can reach
them -- tests/ is not a package, so cross-file fixture imports do not
work. TagCategoriesPanel (shared with a standalone dialog outside the
Hub) gets its roles set on the Hub-wrapped instance only, in
_TagCategoriesPage, so the standalone dialog is untouched.
